adds summaryCards component
